+++ src/DragInput.js
@@ -1,29 +1,29 @@
 import { eventPrefix } from "./settings";
 
 export default class DragInput{
-    static SWIPE_EVENTS = ['mousedown', 'mousemove', 'mouseup', 'touchstart', 'touchmove', 'touchend', 'touchcancel'];
+    // Public
+    _isSwiping = false;
 
-    data;
-    settings;
-    changeFrame;
-    getNextFrame;
-    boundSwipeHandler;
-
-    curX;
-    curY;
-    prevX;
-    prevY;
-    isSwiping = false;
-    threshold;
-    pixelsCorrection = 0;
-    lastInteractionTime;
+    // Internal
+    _curX;
+    _curY;
+    _prevX;
+    _prevY;
+    _threshold;
+    _pixelsCorrection;
+    _lastInteractionTime;
 
     constructor({ data, settings, changeFrame, getNextFrame }) {
-        this.data = data;
-        this.settings = settings;
-        this.changeFrame = changeFrame;
-        this.getNextFrame = getNextFrame;
-        this.boundSwipeHandler = this.#swipeHandler.bind(this);
+        this._data = data;
+        this._settings = settings;
+        this._changeFrame = changeFrame;
+        this._getNextFrame = getNextFrame;
+
+        this._SWIPE_EVENTS = ['mousedown', 'mousemove', 'mouseup', 'touchstart', 'touchmove', 'touchend', 'touchcancel'];
+        this._isSwiping = false;
+        this._boundSwipeHandler = this.#swipeHandler.bind(this);
+        this._pixelsCorrection = 0;
+
         this._updateThreshold();
     }
 
@@ -31,8 +31,8 @@ export default class DragInput{
      * Enable rotating by mouse or touch drag
      */
     _enableDrag(){
-        DragInput.SWIPE_EVENTS.forEach( (value) => {
-            this.data.canvas.element.addEventListener(value, this.boundSwipeHandler);
+        this._SWIPE_EVENTS.forEach( (value) => {
+            this._data.canvas.element.addEventListener(value, this._boundSwipeHandler);
         })
     }
 
@@ -40,13 +40,13 @@ export default class DragInput{
      * Disable rotating by mouse or touch drag
      */
     _disableDrag(){
-        DragInput.SWIPE_EVENTS.forEach( (value) => {
-            this.data.canvas.element.removeEventListener(value, this.boundSwipeHandler);
+        this._SWIPE_EVENTS.forEach( (value) => {
+            this._data.canvas.element.removeEventListener(value, this._boundSwipeHandler);
         })
         // if disabling while swipeMove is running
-        document.removeEventListener('mouseup', this.boundSwipeHandler);
-        document.removeEventListener('mousemove', this.boundSwipeHandler);
-        this.data.canvas.element.style.cursor = null;
+        document.removeEventListener('mouseup', this._boundSwipeHandler);
+        document.removeEventListener('mousemove', this._boundSwipeHandler);
+        this._data.canvas.element.style.cursor = null;
     }
 
     /**
@@ -55,10 +55,10 @@ export default class DragInput{
      */
     _updateThreshold(newValue = null){
         if (newValue) {
-            this.threshold = newValue;
+            this._threshold = newValue;
         }
         else {
-            this.threshold = this.data.canvas.element.clientWidth / this.data.totalImages;
+            this._threshold = this._data.canvas.element.clientWidth / this._data.totalImages;
         }
     }
 
@@ -67,8 +67,8 @@ export default class DragInput{
         // get current click/touch point
         let touches;
         if ( event.touches !== undefined && event.touches.length ) touches = event.touches;
-        this.curX = (touches) ? touches[0].pageX : event.clientX;
-        this.curY = (touches) ? touches[0].pageY : event.clientY;
+        this._curX = (touches) ? touches[0].pageX : event.clientX;
+        this._curY = (touches) ? touches[0].pageY : event.clientY;
 
         switch (event.type){
             case 'mousedown': // start
@@ -77,13 +77,13 @@ export default class DragInput{
                     //event.preventDefault();
                     this.#maybeDisableScroll(event);
                 }
-                document.addEventListener('mouseup', this.boundSwipeHandler); // move outside of the canvas
-                document.addEventListener('mousemove', this.boundSwipeHandler);
+                document.addEventListener('mouseup', this._boundSwipeHandler); // move outside of the canvas
+                document.addEventListener('mousemove', this._boundSwipeHandler);
                 this.#swipeStart();
                 break;
             case 'mousemove':
             case 'touchmove': //move
-                if ( this.isSwiping ) {
+                if ( this._isSwiping ) {
                     //if ( event.type === 'touchmove' && event.cancelable) event.preventDefault();
                     this.#swipeMove();
                 }
@@ -92,67 +92,68 @@ export default class DragInput{
             case 'touchend':
             case 'touchcancel': // end
                 //if ( (event.type === 'touchend' || event.type === 'touchcancel') && event.cancelable) event.preventDefault();
-                if ( this.isSwiping ) {
-                    document.removeEventListener('mouseup', this.boundSwipeHandler);
-                    document.removeEventListener('mousemove', this.boundSwipeHandler);
+                if ( this._isSwiping ) {
+                    document.removeEventListener('mouseup', this._boundSwipeHandler);
+                    document.removeEventListener('mousemove', this._boundSwipeHandler);
                     this.#swipeEnd();
                 }
                 break;
         }
     }
     #swipeStart(){
-        if ( !(this.data.pluginApi.isFastPreloadFinished() || this.data.pluginApi.isPreloadFinished()) ) return;
+        const plugin = this._data.pluginApi;
+        if ( !(plugin.isFastPreloadFinished() || plugin.isPreloadFinished()) ) return;
         // trigger full load after user interaction after fast preload finished
-        if (this.settings.fastPreview && !this.data.pluginApi.isPreloadFinished() && this.data.pluginApi.isFastPreloadFinished()) {
-            this.data.pluginApi.preloadImages();
+        if (this._settings.fastPreview && !plugin.isPreloadFinished() && plugin.isFastPreloadFinished()) {
+            plugin.preloadImages();
         }
-        this.data.pluginApi.stop();
-        this.isSwiping = true;
-        this.data.canvas.element.style.cursor = 'grabbing';
-        this.prevX = this.curX;
-        this.prevY = this.curY;
-        this.data.canvas.element.dispatchEvent( new CustomEvent(eventPrefix + 'drag-start',
-            { detail: {frame: this.data.currentFrame} })
+        plugin.stop();
+        this._isSwiping = true;
+        this._data.canvas.element.style.cursor = 'grabbing';
+        this._prevX = this._curX;
+        this._prevY = this._curY;
+        this._data.canvas.element.dispatchEvent( new CustomEvent(eventPrefix + 'drag-start',
+            { detail: {frame: this._data.currentFrame} })
         );
     }
     #swipeMove(){
         const direction = this.#swipeDirection();
-        const swipeLength = Math.round( Math.abs(this.curX - this.prevX) * this.settings.dragModifier ) + this.pixelsCorrection;
+        const swipeLength = Math.round( Math.abs(this._curX - this._prevX) * this._settings.dragModifier ) + this._pixelsCorrection;
 
-        if ( swipeLength <= this.threshold) return;// Ignore if less than 1 frame
+        if ( swipeLength <= this._threshold) return;// Ignore if less than 1 frame
         if ( direction !== 'left' && direction !== 'right') return; // Ignore vertical directions
 
-        this.prevX = this.curX;
-        this.prevY = this.curY; // Update Y to get right angle
+        this._prevX = this._curX;
+        this._prevY = this._curY; // Update Y to get right angle
 
-        const progress = swipeLength / this.data.canvas.element.width; // full width swipe means full animation
-        let deltaFrames = Math.floor(progress * this.data.totalImages);
-        deltaFrames = deltaFrames % this.data.totalImages;
+        const progress = swipeLength / this._data.canvas.element.width; // full width swipe means full animation
+        let deltaFrames = Math.floor(progress * this._data.totalImages);
+        deltaFrames = deltaFrames % this._data.totalImages;
         // Add pixels to the next swipeMove if frames equivalent of swipe is not an integer number,
         // e.g one frame is 10px, swipeLength is 13px, we change 1 frame and add 3px to the next swipe,
         // so fullwidth swipe is always rotate sprite for 1 turn (with 'dragModifier' = 1)
-        this.pixelsCorrection = swipeLength - (this.threshold * deltaFrames);
+        this._pixelsCorrection = swipeLength - (this._threshold * deltaFrames);
         let isReverse = (direction === 'left'); // left means backward (reverse: true)
-        if (this.settings.inversion) isReverse = !isReverse;// invert direction
-        this.changeFrame(this.getNextFrame( deltaFrames, isReverse )); // left means backward (reverse: true)
-        this.data.canvas.element.dispatchEvent( new CustomEvent(eventPrefix + 'drag-change',
-            { detail: {frame: this.data.currentFrame} })
+        if (this._settings.inversion) isReverse = !isReverse;// invert direction
+        this._changeFrame(this._getNextFrame( deltaFrames, isReverse )); // left means backward (reverse: true)
+        this._data.canvas.element.dispatchEvent( new CustomEvent(eventPrefix + 'drag-change',
+            { detail: {frame: this._data.currentFrame} })
         );
     }
     #swipeEnd(){
         //if ( swipeObject.curX === undefined ) return; // there is no x coord on touch end
-        this.curX = this.curY = this.prevX = this.prevY = null;
-        this.isSwiping = false;
-        this.data.canvas.element.style.cursor = null;
-        this.lastInteractionTime = new Date().getTime();
-        this.data.canvas.element.dispatchEvent( new CustomEvent(eventPrefix + 'drag-end',
-            { detail: {frame: this.data.currentFrame} })
+        this._curX = this._curY = this._prevX = this._prevY = null;
+        this._isSwiping = false;
+        this._data.canvas.element.style.cursor = null;
+        this._lastInteractionTime = new Date().getTime();
+        this._data.canvas.element.dispatchEvent( new CustomEvent(eventPrefix + 'drag-end',
+            { detail: {frame: this._data.currentFrame} })
         );
     }
     #swipeDirection(){
         let xDist, yDist, r, swipeAngle;
-        xDist = this.prevX - this.curX;
-        yDist = this.prevY - this.curY;
+        xDist = this._prevX - this._curX;
+        yDist = this._prevY - this._curY;
         r = Math.atan2(yDist, xDist);
         swipeAngle = Math.round(r * 180 / Math.PI);
         if (swipeAngle < 0) swipeAngle = 360 - Math.abs(swipeAngle);
@@ -169,15 +170,15 @@ export default class DragInput{
      */
     #maybeDisableScroll(event){
         // always prevent
-        if (this.settings.touchScrollMode === "preventPageScroll") event.preventDefault();
+        if (this._settings.touchScrollMode === "preventPageScroll") event.preventDefault();
         // check timer
-        if (this.settings.touchScrollMode === "pageScrollTimer") {
-            let now = new Date().getTime();
+        if (this._settings.touchScrollMode === "pageScrollTimer") {
+            const now = new Date().getTime();
             // less time than delay => prevent page scroll
-            if (this.lastInteractionTime && (now - this.lastInteractionTime < this.settings.pageScrollTimerDelay) ){
+            if (this._lastInteractionTime && (now - this._lastInteractionTime < this._settings.pageScrollTimerDelay) ){
                 event.preventDefault();
             } else { // more time than delay or first interaction => clear timer
-                this.lastInteractionTime = null;
+                this._lastInteractionTime = null;
             }
         }
         // if touchScrollMode="allowPageScroll" => don't prevent scroll