+++ src/Poster.js
@@ -1,27 +1,26 @@
 import { eventPrefix } from "./settings";
 
 export default class Poster{
-    settings;
-    data;
-    drawFrame;
-
-    imageObject = null;
-    isPosterLoaded = false;
+    // Internal
+    _imageObject;
+    _isPosterLoaded;
 
     constructor({settings, data, drawFrame}) {
-        this.settings = settings;
-        this.data = data;
-        this.drawFrame = drawFrame;
+        this._settings = settings;
+        this._data = data;
+        this._drawFrame = drawFrame;
+
+        this._isPosterLoaded = false;
     }
 
     /**
      * Start loading poster, then  show if needed
      */
     _loadAndShowPoster(){
-        if (this.settings.poster && !this.data.isAnyFrameChanged) {
-            this.imageObject = new Image();
-            this.imageObject.onload = this.imageObject.onerror = this.#onPosterLoaded.bind(this);
-            this.imageObject.src = this.settings.poster;
+        if (this._settings.poster && !this._data.isAnyFrameChanged) {
+            this._imageObject = new Image();
+            this._imageObject.onload = this._imageObject.onerror = this.#onPosterLoaded.bind(this);
+            this._imageObject.src = this._settings.poster;
         }
     }
 
@@ -29,23 +28,23 @@ export default class Poster{
      * Redraw poster after canvas change if the poster was loaded
      */
     _redrawPoster(){
-        if ( this.data.isAnyFrameChanged || !this.isPosterLoaded ) return;
+        if ( this._data.isAnyFrameChanged || !this._isPosterLoaded ) return;
         this.#drawPoster();
     }
 
     #onPosterLoaded(e){
         if (e.type === "error") return;
-        this.isPosterLoaded = true;
-        this.data.canvas.element.dispatchEvent( new Event(eventPrefix + 'poster-loaded') );
-        this.settings.onPosterLoaded(this.data.pluginApi);
+        this._isPosterLoaded = true;
+        this._data.canvas.element.dispatchEvent( new Event(eventPrefix + 'poster-loaded') );
+        this._settings.onPosterLoaded(this._data.pluginApi);
         // show only if there wasn't any frame change from initial
         // if poster loaded after all the images and any action, it won't be shown
-        if ( !this.data.isAnyFrameChanged ) {
+        if ( !this._data.isAnyFrameChanged ) {
             this.#drawPoster();
         }
     }
 
     #drawPoster(){
-        this.drawFrame(this.imageObject, { settings: this.settings, data: this.data });
+        this._drawFrame(this._imageObject);
     }
 }